+++ peer/impl/message/msghandler.go
@@ -1,0 +1,53 @@
+package message
+
+import (
+	"github.com/rs/zerolog/log"
+	"go.dedis.ch/cs438/transport"
+	"go.dedis.ch/cs438/types"
+	"golang.org/x/xerrors"
+)
+
+// ProcessChatMessage is a callback function to handle received chat message
+func (m *MessageModule) ProcessChatMessage(msg types.Message, pkt transport.Packet) error {
+	// cast the message to its actual type. You assume it is the right type.
+	chatMsg, ok := msg.(*types.ChatMessage)
+	if !ok {
+		return xerrors.Errorf("wrong type: %T", msg)
+	}
+
+	log.Info().Msgf("%s received a chat message from: %s. Msg: %s",
+		m.conf.Socket.GetAddress(),
+		pkt.Header.Source,
+		chatMsg.Message)
+	return nil
+}
+
+// ProcessPrivateMsg is a callback function to handle received private message
+func (m *MessageModule) ProcessPrivateMsg(msg types.Message, pkt transport.Packet) error {
+	privateMsg, ok := msg.(*types.PrivateMessage)
+	if !ok {
+		return xerrors.Errorf("wrong type: %T", msg)
+	}
+
+	_, ok = privateMsg.Recipients[m.conf.Socket.GetAddress()]
+	if ok {
+		// process the message if in the recipients list
+		newPkt := transport.Packet{
+			Header: pkt.Header,
+			Msg:    privateMsg.Msg,
+		}
+		err := m.conf.MessageRegistry.ProcessPacket(newPkt)
+		return err
+	}
+	return nil
+}
+
+// ProcessEmptyMsg takes the empty message and do nothing
+func (m *MessageModule) ProcessEmptyMsg(msg types.Message, pkt transport.Packet) error {
+	_, ok := msg.(*types.EmptyMessage)
+	if !ok {
+		return xerrors.Errorf("wrong type: %T", msg)
+	}
+
+	return nil
+}
